+++ src/main/java/no/vegvesen/nvdbapi/client/gson/RoadSysRefParser.java
@@ -26,6 +26,7 @@
 package no.vegvesen.nvdbapi.client.gson;
 
 import com.google.gson.JsonObject;
+import no.vegvesen.nvdbapi.client.model.Direction;
 import no.vegvesen.nvdbapi.client.model.roadnet.roadsysref.*;
 
 import static java.util.Objects.isNull;
@@ -85,7 +86,7 @@ public final class RoadSysRefParser {
             parseStringMember(obj, "trafikantgruppe"),
             getFromMeter(obj),
             getToMeter(obj),
-            parseStringMember(obj, "retning"));
+            Direction.from(parseStringMember(obj, "retning")));
     }
 
     private static Intersection parseIntersection(JsonObject obj) {
@@ -100,7 +101,7 @@ public final class RoadSysRefParser {
                 parseIntMember(obj, "kryssdel"),
                 getFromMeter(obj),
                 getToMeter(obj),
-                parseStringMember(obj, "retning"));
+                Direction.from(parseStringMember(obj, "retning")));
         }
         return null;
     }
@@ -117,7 +118,7 @@ public final class RoadSysRefParser {
                 parseIntMember(obj, "sideanleggsdel"),
                 getFromMeter(obj),
                 getToMeter(obj),
-                parseStringMember(obj, "retning"));
+                Direction.from(parseStringMember(obj, "retning")));
         }
         return null;
     }